move CLI error handling to strata-cli-common crate
- update alpen-cli to use strata-cli-common for errors
- update strata-dbtool to use strata-cli-common for errors
