+++ Cargo.toml
@@ -74,6 +74,7 @@ alpen-reth-primitives = { path = "crates/reth/primitives" }
 alpen-reth-rpc = { path = "crates/reth/rpc" }
 strata-btcio = { path = "crates/btcio" }
 strata-chaintsn = { path = "crates/chaintsn" }
+strata-cli-common = { path = "crates/strata-cli-common" }
 strata-common = { path = "crates/common" }
 strata-config = { path = "crates/config" }
 strata-consensus-logic = { path = "crates/consensus-logic" }
@@ -196,7 +197,7 @@ borsh = { version = "1.5.0", features = ["derive"] }
 bytes = "1.6.0"
 cfg-if = "1.0.0"
 chrono = "0.4.38"
-clap = "4"
+clap = { version="4.5.39", default-features = false, features = ["derive"] }
 deadpool = "0.12.1"
 digest = "0.10"
 ethnum = "1.5.0"
