+++ bin/alpen-cli/Cargo.toml
@@ -8,6 +8,7 @@ name = "alpen"
 path = "src/main.rs"
 
 [dependencies]
+strata-cli-common.workspace = true
 strata-primitives.workspace = true
 
 aes-gcm-siv = "0.11.1"
