+++ bin/alpen-cli/src/seed/keychain.rs
@@ -73,10 +73,43 @@ impl EncryptedSeedPersister for KeychainPersister {
     }
 }
 
-use crate::errors::{NoStorageAccess, PlatformFailure};
+use crate::signet::backend::BoxedErr;
 
 // below is wrapper around [`keyring::Error`] so it can be used with OneOf to more precisely handle
 // errors
+
+/// This indicates runtime failure in the underlying platform storage system. The details of the
+/// failure can be retrieved from the attached platform error.
+#[derive(Debug)]
+#[allow(unused)]
+pub struct PlatformFailure(BoxedErr);
+
+#[allow(unused)]
+impl PlatformFailure {
+    pub fn new<E>(e: E) -> Self
+    where
+        E: Into<BoxedErr>,
+    {
+        Self(e.into())
+    }
+}
+
+/// This indicates that the underlying secure storage holding saved items could not be accessed.
+/// Typically this is because of access rules in the platform; for example, it might be that the
+/// credential store is locked. The underlying platform error will typically give the reason.
+#[derive(Debug)]
+#[allow(unused)]
+pub struct NoStorageAccess(BoxedErr);
+
+#[allow(unused)]
+impl NoStorageAccess {
+    pub fn new<E>(e: E) -> Self
+    where
+        E: Into<BoxedErr>,
+    {
+        Self(e.into())
+    }
+}
 
 /// This indicates that there is no underlying credential entry in the platform for this entry.
 /// Either one was never set, or it was deleted.
