+++ bin/strata-dbtool/Cargo.toml
@@ -12,5 +12,6 @@ argh.workspace = true
 rockbound.workspace = true
 tracing-subscriber.workspace = true
 
+strata-cli-common.workspace = true
 strata-db.workspace = true
 strata-rocksdb.workspace = true
