+++ bin/strata-dbtool/src/cmd/client_state.rs
@@ -1,4 +1,5 @@
 use argh::FromArgs;
+use strata_cli_common::errors::DisplayedError;
 use strata_db::traits::Database;
 
 use crate::cli::OutputFormat;
@@ -19,6 +20,6 @@ pub(crate) struct GetClientStateUpdateArgs {
 pub(crate) fn get_client_state_update(
     _db: &impl Database,
     _args: GetClientStateUpdateArgs,
-) -> Result<(), Box<dyn std::error::Error>> {
+) -> Result<(), DisplayedError> {
     Ok(())
 }
