+++ bin/strata-dbtool/src/main.rs
