+++ crates/strata-cli-common/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+edition = "2021"
+name = "strata-cli-common"
+version = "0.3.0-alpha.1"
