+++ crates/strata-cli-common/src/lib.rs
@@ -1,0 +1,2 @@
+//! Reusable error types for CLI tools.
+pub mod errors;
